requirements.txt
langchain
langgraph
playwright
langchain-fireworks
langchain_openai
bs4
python-dotenv
pydantic